malseq/requirements.txt
torch>=2.2
torchvision
torchaudio
numpy
scikit-learn
pandas
scipy
pyyaml
tqdm
pefile
lief
capstone
einops
matplotlib
seaborn
scikit-multiflow
